Regularise according to norm of vectors

+++ wallhack/rankingexp/RegularisationExp.py
@@ -3,7 +3,9 @@ import logging
 import sys
 import os
 import multiprocessing
+import itertools 
 from sandbox.recommendation.MaxLocalAUC import MaxLocalAUC
+from sandbox.recommendation.IterativeSoftImpute import IterativeSoftImpute
 from sandbox.util.SparseUtils import SparseUtils
 from sandbox.util.MCEvaluator import MCEvaluator
 from sandbox.util.PathDefaults import PathDefaults
@@ -11,7 +13,8 @@ from sandbox.util.Sampling import Sampling
 from wallhack.rankingexp.DatasetUtils import DatasetUtils
 
 """
-We look at the ROC curves on the test set for different values of lambda.  
+We look at the ROC curves on the test set for different values of maxNorm. We want 
+to find why on epinions, the learning overfits so vary lambdaU and lambdaV  
 """
 
 
@@ -23,12 +26,11 @@ numpy.seterr(all="raise")
 if len(sys.argv) > 1:
     dataset = sys.argv[1]
 else: 
-    dataset = "synthetic2"
+    dataset = "synthetic"
 
 saveResults = True
 prefix = "Regularisation"
 outputFile = PathDefaults.getOutputDir() + "ranking/" + prefix + dataset.title() + "Results.npz" 
-print(outputFile)
 
 if dataset == "synthetic": 
     X, U, V = DatasetUtils.syntheticDataset1()
@@ -36,9 +38,12 @@ elif dataset == "synthetic2":
     X = DatasetUtils.syntheticDataset2()
 elif dataset == "movielens": 
     X = DatasetUtils.movieLens()
+elif dataset == "epinions": 
+    X = DatasetUtils.epinions()
+    X, userInds = Sampling.sampleUsers2(X, 10000)    
 elif dataset == "flixster": 
     X = DatasetUtils.flixster()
-    X, userInds = Sampling.sampleUsers2(X, 50000)
+    X, userInds = Sampling.sampleUsers2(X, 10000)
 else: 
     raise ValueError("Unknown dataset: " + dataset)
 
@@ -46,28 +51,31 @@ m, n = X.shape
 u = 0.1 
 w = 1-u
 
+logging.debug("Sampled X shape: " + str(X.shape))
+
 testSize = 5
 folds = 5
 trainTestXs = Sampling.shuffleSplitRows(X, folds, testSize)
 
 numRecordAucSamples = 200
 
-k2 = 8
+k2 = 64
 u2 = 0.5
 w2 = 1-u2
 eps = 10**-8
-lmbda = 0.1
+lmbda = 0.0
 maxLocalAuc = MaxLocalAUC(k2, w2, eps=eps, lmbdaU=lmbda, lmbdaV=lmbda, stochastic=True)
 maxLocalAuc.alpha = 0.1
-maxLocalAuc.alphas = 2.0**-numpy.arange(0, 5, 1)
+maxLocalAuc.alphas = 2.0**-numpy.arange(0, 9, 1)
 maxLocalAuc.folds = 1
 maxLocalAuc.initialAlg = "rand"
 maxLocalAuc.itemExpP = 0.0
 maxLocalAuc.itemExpQ = 0.0
 maxLocalAuc.ks = numpy.array([k2])
-maxLocalAuc.lmbdas = 2.0**-numpy.arange(0, 6)
+maxLocalAuc.lmbdas = 2.0**-numpy.arange(-5, 6, 3)
 maxLocalAuc.loss = "hinge"
 maxLocalAuc.maxIterations = 100
+maxLocalAuc.maxNorm = numpy.sqrt(0.5)
 maxLocalAuc.metric = "f1"
 maxLocalAuc.normalise = True
 maxLocalAuc.numAucSamples = 10
@@ -82,6 +90,12 @@ maxLocalAuc.t0s = 2.0**-numpy.arange(7, 12, 1)
 maxLocalAuc.validationSize = 3
 maxLocalAuc.validationUsers = 0
 
+softImpute = IterativeSoftImpute(k=k2, postProcess=True)
+
+maxNorms = 2.0**numpy.arange(-1, 6)
+
+#numProcesses = 1
+numProcesses = multiprocessing.cpu_count()
 os.system('taskset -p 0xffffffff %d' % os.getpid())
 
 logging.debug("Starting training")
@@ -95,7 +109,7 @@ def computeTestAuc(args):
     U, V, trainMeasures, testMeasures, iterations, time = maxLocalAuc.learnModel(trainX, U=U, V=V, verbose=True)
     
     fprTrain, tprTrain = MCEvaluator.averageRocCurve(trainX, U, V)
-    fprTest, tprTest = MCEvaluator.averageRocCurve(testX, U, V)
+    fprTest, tprTest = MCEvaluator.averageRocCurve(testX, U, V, trainX=trainX)
         
     return fprTrain, tprTrain, fprTest, tprTest
 
@@ -103,27 +117,42 @@ if saveResults:
     paramList = []
     chunkSize = 1
     
+    #First generate SoftImpute results as a benchmark. 
+    trainX, testX = trainTestXs[0]
+    learner = softImpute.copy()
+    
+    trainIterator = iter([trainX.toScipyCsc()])
+    ZList = learner.learnModel(trainIterator)    
+    U, s, V = ZList.next()
+    U = U*s
+    
+    U = numpy.ascontiguousarray(U)
+    V = numpy.ascontiguousarray(V)
+    
+    fprTrainSI, tprTrainSI = MCEvaluator.averageRocCurve(trainX, U, V)
+    fprTestSI, tprTestSI = MCEvaluator.averageRocCurve(testX, U, V, trainX=trainX)
+    
+    #Now train MaxLocalAUC 
     U, V = maxLocalAuc.initUV(X)
     
-    for lmbda in maxLocalAuc.lmbdas: 
+    for maxNorm in maxNorms:  
         for trainX, testX in trainTestXs: 
             learner = maxLocalAuc.copy()
-            learner.lmbdaU = lmbda 
-            learner.lmbdaV = lmbda 
+            learner.maxNorm = maxNorm 
             paramList.append((trainX, testX, learner, U.copy(), V.copy()))
 
-    pool = multiprocessing.Pool(maxtasksperchild=100, processes=multiprocessing.cpu_count())
-    resultsIterator = pool.imap(computeTestAuc, paramList, chunkSize)
-    
-    #import itertools 
-    #resultsIterator = itertools.imap(computeTestAuc, paramList)
+    if numProcesses != 1: 
+        pool = multiprocessing.Pool(maxtasksperchild=100, processes=multiprocessing.cpu_count())
+        resultsIterator = pool.imap(computeTestAuc, paramList, chunkSize)
+    else: 
+        resultsIterator = itertools.imap(computeTestAuc, paramList)
     
     meanFprTrains = []
     meanTprTrains = []
     meanFprTests = []
     meanTprTests = []
     
-    for lmbda in maxLocalAuc.lmbdas: 
+    for maxNorm in maxNorms:  
         fprTrains = [] 
         tprTrains = [] 
         fprTests = [] 
@@ -147,42 +176,55 @@ if saveResults:
         meanFprTests.append(meanFprTest)
         meanTprTests.append(meanTprTest)
         
-    numpy.savez(outputFile, meanFprTrains, meanTprTrains, meanFprTests, meanTprTests)
+    numpy.savez(outputFile, meanFprTrains, meanTprTrains, meanFprTests, meanTprTests, fprTrainSI, tprTrainSI, fprTestSI, tprTestSI)
     
-    pool.terminate()   
+    if numProcesses != 1: 
+        pool.terminate()   
     logging.debug("Saved results in " + outputFile)
 else: 
     data = numpy.load(outputFile)
-    meanFprTrain, meanTprTrain, meanFprTest, meanTprTest = data["arr_0"], data["arr_1"], data["arr_2"], data["arr_3"]      
+    meanFprTrain, meanTprTrain, meanFprTest, meanTprTest, fprTrainSI, tprTrainSI, fprTestSI, tprTestSI = data["arr_0"], data["arr_1"], data["arr_2"], data["arr_3"], data["arr_4"], data["arr_5"], data["arr_6"], data["arr_7"]      
    
     import matplotlib 
     matplotlib.use("GTK3Agg")
     import matplotlib.pyplot as plt   
     
-    plotInds = ["k-", "k--", "k-.", "r-", "b-", "c-", "c--", "c-.", "g-", "g--", "g-."]
+    #print(meanFprTrain[0, :])
+    #print(meanTprTrain[0, :])
     
-    for i, lmbda in enumerate(maxLocalAuc.lmbdas):
+    plotInds = ["k-", "k--", "k-.", "k:", "r-", "r--", "r-.", "r:", "g-", "g--", "g-.", "g:", "b-", "b--", "b-.", "b:", "c-"]
+    ind = 0 
+    
+    for i, maxNorm in enumerate(maxNorms):
+        label = r"$maxNorm=$" + str(maxNorm)
 
-        label = r"$\lambda=$" + str(lmbda)
-
+        fprTrainStart =   meanFprTrain[ind, meanFprTrain[ind, :]<=0.2]   
+        tprTrainStart =   meanTprTrain[ind, meanFprTrain[ind, :]<=0.2]
         
-        fprTrainStart =   meanFprTrain[i, meanFprTrain[i, :]<=0.2]   
-        tprTrainStart =   meanTprTrain[i, meanFprTrain[i, :]<=0.2]   
+        print(fprTrainStart, tprTrainStart)
         
         plt.figure(0)
-        plt.plot(fprTrainStart, tprTrainStart, plotInds[i], label=label)
+        plt.plot(fprTrainStart, tprTrainStart, plotInds[ind], label=label)
         
         plt.figure(1)
-        plt.plot(meanFprTrain[i, :], meanTprTrain[i, :], plotInds[i], label=label)
+        plt.plot(meanFprTrain[ind, :], meanTprTrain[ind, :], plotInds[ind], label=label)
         
-        fprTestStart =   meanFprTest[i, meanFprTest[i, :]<=0.2]   
-        tprTestStart =   meanTprTest[i, meanFprTest[i, :]<=0.2]         
+        fprTestStart =   meanFprTest[ind, meanFprTest[ind, :]<=0.2]   
+        tprTestStart =   meanTprTest[ind, meanFprTest[ind, :]<=0.2]         
         
         plt.figure(2)    
-        plt.plot(fprTestStart, tprTestStart, plotInds[i], label=label)            
+        plt.plot(fprTestStart, tprTestStart, plotInds[ind], label=label)            
         
         plt.figure(3)    
-        plt.plot(meanFprTest[i, :], meanTprTest[i, :], plotInds[i], label=label)    
+        plt.plot(meanFprTest[ind, :], meanTprTest[ind, :], plotInds[ind], label=label)    
+        
+        ind += 1
+    
+    plt.figure(1)
+    plt.plot(fprTrainSI, tprTrainSI, plotInds[ind], label="SI")    
+    
+    plt.figure(3)
+    plt.plot(fprTestSI, tprTestSI , plotInds[ind], label="SI")       
     
     plt.figure(0)
     plt.xlabel("false positive rate")
